Module damn_vuln_blockchain::asset

Assets are objects that can be transacted on the blockchain

The easiest way to manage Asset with te AssetLedger actor

AssetLedger Messages:

ChangeAssetOwner: Changes an asset's owner

InitNetwork: Initializes assets to peers in the network

GetAssetInfo: Get an asset's info

DumpLedger: Dump the entire asset ledger

ReplaceLedger: Replace the current ledger with another ledger, useful when

ChooseValidator: Choose validator based on coinage

GetPeerAssets: Get all the assets belonging to a peer synchronising state

Structs

Asset
/// Assets are objects that can be transacted on the blockchain

AssetLedger
represents the world(full network) state of Asset

ChangeAssetOwner
Changes owner of asset ChangeOwner.0 to ChangeOwner.1

ChangeAssetOwnerBuilder
Builder for ChangeAssetOwner.

ChooseValidator
Get asset info of GetAssetInfo.0

DumpLedger
Dumps entire ledger

GetAssetInfo
Get asset info of GetAssetInfo.0

GetPeerAssets
Get assets belonging to a peer

InitNetwork
Initializes assets to peers in the network AssetLedger` will automatically devide available assets please note that this call will result in error if it's called after all the assets' owners are changed from None to Some(val)

InitNetworkBuilder
Builder for InitNetwork.

ReplaceLedger
Replaces asset ledger Useful when forking