Module damn_vuln_blockchain::discovery

Peer management service for synchronising and discovery. The easiest way to interact with the peer management service is with the Network actor.

Network actor supports the following messages:

AddPeer: Add a peer to the network

DumpPeer: Get a dump of all peers in the network

GetPeer: Get ingo on a specific peer

Structs

AddPeer
Add peer

DumpPeer
Dump all peers

GetCurrentSize
Get current network size

GetPeer
Get peer of ID

Network